pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "odoo-extractor"
version = "0.1.0"
description = "Extrator de dados do Odoo usando XML-RPC com exportação para Parquet"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Seu Nome", email = "seu-email@example.com"}
]
keywords = ["odoo", "extractor", "xmlrpc", "polars", "parquet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.3.4",
    "polars>=1.35.1",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
odoo-extractor = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

